Adding everything into one project
